Fold DocumentSpecificOptionSet into DocumentOptionSet
