+++ src/Workspaces/Core/Portable/Options/DocumentOptionSet.cs
@@ -2,9 +2,14 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Diagnostics;
+using System.Linq;
 using Microsoft.CodeAnalysis.Diagnostics;
+using Microsoft.CodeAnalysis.ErrorReporting;
+using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.Options
 {
@@ -15,28 +20,79 @@ namespace Microsoft.CodeAnalysis.Options
     /// </summary>
     public sealed class DocumentOptionSet : OptionSet
     {
-        private readonly OptionSet _backingOptionSet;
+        private readonly OptionSet _underlyingOptions;
+        private readonly StructuredAnalyzerConfigOptions? _configOptions;
+        private ImmutableDictionary<OptionKey, object?> _values;
         private readonly string _language;
 
-        internal DocumentOptionSet(OptionSet backingOptionSet, string language)
+        internal DocumentOptionSet(StructuredAnalyzerConfigOptions? configOptions, OptionSet underlyingOptions, string language)
+            : this(configOptions, underlyingOptions, language, ImmutableDictionary<OptionKey, object?>.Empty)
         {
-            _backingOptionSet = backingOptionSet;
+        }
+
+        internal DocumentOptionSet(StructuredAnalyzerConfigOptions? configOptions, OptionSet underlyingOptions, string language, ImmutableDictionary<OptionKey, object?> values)
+        {
             _language = language;
+            _configOptions = configOptions;
+            _underlyingOptions = underlyingOptions;
+            _values = values;
         }
 
         internal string Language => _language;
 
+        [PerformanceSensitive("https://github.com/dotnet/roslyn/issues/30819", AllowLocks = false)]
         private protected override object? GetOptionCore(OptionKey optionKey)
-            => _backingOptionSet.GetOption(optionKey);
+        {
+            // If we already know the document specific value, we're done
+            if (_values.TryGetValue(optionKey, out var value))
+            {
+                return value;
+            }
+
+            if (TryGetAnalyzerConfigOption(optionKey, out value))
+            {
+                // Cache and return
+                return ImmutableInterlocked.GetOrAdd(ref _values, optionKey, value);
+            }
+
+            // We don't have a document specific value, so forward
+            return _underlyingOptions.GetOption(optionKey);
+        }
+
+        private bool TryGetAnalyzerConfigOption(OptionKey option, out object? value)
+        {
+            if (_configOptions == null)
+            {
+                value = null;
+                return false;
+            }
+
+            var editorConfigPersistence = (IEditorConfigStorageLocation?)option.Option.StorageLocations.SingleOrDefault(static location => location is IEditorConfigStorageLocation);
+            if (editorConfigPersistence == null)
+            {
+                value = null;
+                return false;
+            }
+
+            try
+            {
+                return editorConfigPersistence.TryGetOption(_configOptions, option.Option.Type, out value);
+            }
+            catch (Exception e) when (FatalError.ReportAndCatch(e))
+            {
+                value = null;
+                return false;
+            }
+        }
 
         public T GetOption<T>(PerLanguageOption<T> option)
-            => _backingOptionSet.GetOption(option, _language);
+            => GetOption(option, _language);
 
         internal T GetOption<T>(PerLanguageOption2<T> option)
-            => _backingOptionSet.GetOption(option, _language);
+            => GetOption(option, _language);
 
         public override OptionSet WithChangedOption(OptionKey optionAndLanguage, object? value)
-            => new DocumentOptionSet(_backingOptionSet.WithChangedOption(optionAndLanguage, value), _language);
+            => new DocumentOptionSet(_configOptions, _underlyingOptions, _language, _values.SetItem(optionAndLanguage, value));
 
         /// <summary>
         /// Creates a new <see cref="DocumentOptionSet" /> that contains the changed value.
@@ -53,10 +109,10 @@ namespace Microsoft.CodeAnalysis.Options
         private protected override AnalyzerConfigOptions CreateAnalyzerConfigOptions(IOptionService optionService, string? language)
         {
             Debug.Assert((language ?? _language) == _language, $"Use of a {nameof(DocumentOptionSet)} is not expected to differ from the language it was constructed with.");
-            return _backingOptionSet.AsAnalyzerConfigOptions(optionService, language ?? _language);
+            return base.CreateAnalyzerConfigOptions(optionService, language ?? _language);
         }
 
         internal override IEnumerable<OptionKey> GetChangedOptions(OptionSet optionSet)
-            => _backingOptionSet.GetChangedOptions(optionSet);
+            => GetChangedOptions(optionSet);
     }
 }
