+++ src/Workspaces/Core/Portable/Options/OptionServiceFactory.cs
@@ -126,87 +126,6 @@ namespace Microsoft.CodeAnalysis.Options
             public void SetOptions(OptionSet optionSet) => _globalOptionService.SetOptions(optionSet);
             public void RegisterWorkspace(Workspace workspace) => _globalOptionService.RegisterWorkspace(workspace);
             public void UnregisterWorkspace(Workspace workspace) => _globalOptionService.UnregisterWorkspace(workspace);
-
-            public async Task<OptionSet> GetUpdatedOptionSetForDocumentAsync(Document document, OptionSet optionSet, CancellationToken cancellationToken)
-            {
-                var provider = (ProjectState.ProjectAnalyzerConfigOptionsProvider)document.Project.State.AnalyzerOptions.AnalyzerConfigOptionsProvider;
-                var options = await provider.GetOptionsAsync(document.DocumentState, cancellationToken).ConfigureAwait(false);
-                return new DocumentSpecificOptionSet(options, optionSet);
-            }
-
-            private sealed class DocumentSpecificOptionSet : OptionSet
-            {
-                private readonly OptionSet _underlyingOptions;
-                private readonly StructuredAnalyzerConfigOptions? _configOptions;
-                private ImmutableDictionary<OptionKey, object?> _values;
-
-                public DocumentSpecificOptionSet(StructuredAnalyzerConfigOptions? configOptions, OptionSet underlyingOptions)
-                    : this(configOptions, underlyingOptions, ImmutableDictionary<OptionKey, object?>.Empty)
-                {
-                }
-
-                public DocumentSpecificOptionSet(StructuredAnalyzerConfigOptions? configOptions, OptionSet underlyingOptions, ImmutableDictionary<OptionKey, object?> values)
-                {
-                    _configOptions = configOptions;
-                    _underlyingOptions = underlyingOptions;
-                    _values = values;
-                }
-
-                [PerformanceSensitive("https://github.com/dotnet/roslyn/issues/30819", AllowLocks = false)]
-                private protected override object? GetOptionCore(OptionKey optionKey)
-                {
-                    // If we already know the document specific value, we're done
-                    if (_values.TryGetValue(optionKey, out var value))
-                    {
-                        return value;
-                    }
-
-                    if (TryGetAnalyzerConfigOption(optionKey, out value))
-                    {
-                        // Cache and return
-                        return ImmutableInterlocked.GetOrAdd(ref _values, optionKey, value);
-                    }
-
-                    // We don't have a document specific value, so forward
-                    return _underlyingOptions.GetOption(optionKey);
-                }
-
-                private bool TryGetAnalyzerConfigOption(OptionKey option, out object? value)
-                {
-                    if (_configOptions == null)
-                    {
-                        value = null;
-                        return false;
-                    }
-
-                    var editorConfigPersistence = (IEditorConfigStorageLocation?)option.Option.StorageLocations.SingleOrDefault(static location => location is IEditorConfigStorageLocation);
-                    if (editorConfigPersistence == null)
-                    {
-                        value = null;
-                        return false;
-                    }
-
-                    try
-                    {
-                        return editorConfigPersistence.TryGetOption(_configOptions, option.Option.Type, out value);
-                    }
-                    catch (Exception e) when (FatalError.ReportAndCatch(e))
-                    {
-                        value = null;
-                        return false;
-                    }
-                }
-
-                public override OptionSet WithChangedOption(OptionKey optionAndLanguage, object? value)
-                    => new DocumentSpecificOptionSet(_configOptions, _underlyingOptions, _values.SetItem(optionAndLanguage, value));
-
-                internal override IEnumerable<OptionKey> GetChangedOptions(OptionSet optionSet)
-                {
-                    // GetChangedOptions only needs to be supported for OptionSets that need to be compared during application,
-                    // but that's already enforced it must be a full SerializableOptionSet.
-                    throw new NotSupportedException();
-                }
-            }
         }
     }
 }
